fixed dummy model func
